Structured usecases into more readable format. Calculating functionality at the end of quiz screen + saving thos points in database. Fixed bug with no selected option in last question. Deleted answered question argument in GameOver Screen

+++ app/src/main/java/com/jakubn/codequizapp/domain/model/User.kt
@@ -14,5 +14,5 @@ data class User(
     val gamesPlayed: Int = 0,
     val wins: Int = 0,
     val score: Int = 0,
-    val winRatio: Int = 0
+    val winRatio: Float = 0f
 ): Parcelable
